@@ -1,20 +1,43 @@
 {
-  "name": "facerecognition",
+  "name": "smart-brain-front",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "clarifai": "^2.8.0",
-    "react": "^16.4.2",
-    "react-dom": "^16.4.2",
-    "react-particles-js": "^2.3.0",
-    "react-scripts": "1.1.4",
-    "react-tilt": "^0.1.4",
-    "tachyons": "^4.11.1"
+    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/react": "^11.1.0",
+    "@testing-library/user-event": "^12.1.10",
+    "node-sass": "^5.0.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-particles-js": "^3.4.1",
+    "react-scripts": "4.0.3",
+    "react-spring": "^9.1.0",
+    "react-use-gesture": "^9.1.3",
+    "tachyons": "^4.12.0",
+    "web-vitals": "^1.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
