@@ -1,25 +1,20 @@
-import React from 'react';
-import './ImageLinkForm.css';
+
 
 const ImageLinkForm = ({ onInputChange, onPictureSubmit }) => {
-  return(
+  return (
     <div>
-      <p className='f3 mb3'>
-        {'The All Seeing Eye can recognize faces. Give it a whirl.'}
-      </p>
-      <div className="center">
-        <div className='form center pa4 br3 shadow-5'>
-          <input
-            className="f4 pa2 w-70 center"
-            type="text"
-            placeholder="https://"
-            onChange={onInputChange}
-          />
-          <button className="w-30 grow f4 link ph3 pv2 diib" onClickCapture={onPictureSubmit}>Detect</button>
+      <p className="f3 mb0 white tc">I can see your face.</p>
+      <div className="flex justify-center">
+        <div className="w-50-l pa4 br3 flex flex-row-ns flex-column items-center">
+          <input className="f4 pa2 w-70 center" type="text" name="" id="" onChange={onInputChange} />
+          <button
+            className="w-70 w-30-ns f4 ph3 pv2 dib bg-white black bn dim"
+            onClick={onPictureSubmit}
+          > Detect </button>
         </div>
       </div>
     </div>
-  );
+  )
 }
 
 export default ImageLinkForm;
