@@ -1,20 +1,16 @@
-import React from 'react';
 
 const Navigation = ({ onRouteChange, isSignedIn }) => {
-    if(isSignedIn) {
-      return (
-        <nav style={{display: 'flex', justifyContent: 'flex-end'}}>
-          <p onClickCapture={() => onRouteChange('signin')} className='f3 link dim black underline pa3 pointer'>Sign Out</p>
-        </nav>
-      );
-    } else {
-      return (
-        <nav style={{display: 'flex', justifyContent: 'flex-end'}}>
-          <p onClickCapture={() => onRouteChange('signin')} className='f4 link dim black underline pa3 pointer'>Sign In</p>
-          <p onClickCapture={() => onRouteChange('register')} className='f4 link dim black underline pa3 pointer'>Register</p>
-        </nav>
-      );
-    }
+  return (
+    isSignedIn ?
+      <nav className="absolute right-1">
+        <p onClick={() => onRouteChange('signin')} className="f3 link dim black underline pa3 pointer">Sign Out</p>
+      </nav>
+    :
+      <nav className="absolute right-1">
+        {/* <p onClick={() => onRouteChange('signup')} className="f3 link dim black underline pa3 pointer">Sign Up</p> */}
+        {/* <p onClick={() => onRouteChange('signin')} className="f3 link dim black underline pa3 pointer">Sign In</p> */}
+      </nav>
+  )
 }
 
 export default Navigation;
