@@ -1,13 +1,11 @@
-import React from 'react';
-
-const Rank = ({ name, entries }) => {
-  return(
-    <div className='mb3'>
-      <div className='white f3'>
-        {`${name}, your current entry count is`}
+const Rank = ({ user }) => {
+  return (
+    <div className="white tc">
+      <div className="f3">
+        { `${user.first_name} ${user?.last_name}` || 'Hey there' }, your current rank is
       </div>
-      <div className='white f1'>
-        {entries}
+      <div className="f1">
+        { user.entries || 0 }
       </div>
     </div>
   );
